cinebook-main/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
celery>=5.3
redis>=5.0
psycopg2-binary>=2.9
python-decouple>=3.8
Pillow>=10.0
django-filter>=23.0
gunicorn>=21.2
whitenoise>=6.6
dj-database-url>=2.1
